Move traceHandler.go into a separate folder
