+++ pkg/server/query/entryHandlers.go
@@ -35,7 +35,7 @@ import (
 	prom "github.com/siglens/siglens/pkg/integrations/prometheus/promql"
 	"github.com/siglens/siglens/pkg/querytracker"
 	"github.com/siglens/siglens/pkg/sampledataset"
-	tracing "github.com/siglens/siglens/pkg/segment/tracing"
+	tracinghandler "github.com/siglens/siglens/pkg/segment/tracing/handler"
 	usq "github.com/siglens/siglens/pkg/usersavedqueries"
 	log "github.com/sirupsen/logrus"
 	"github.com/valyala/fasthttp"
@@ -475,6 +475,6 @@ func liveTailHandler(myid uint64) func(ctx *fasthttp.RequestCtx) {
 // Tracing apis
 func searchTracesHandler() func(ctx *fasthttp.RequestCtx) {
 	return func(ctx *fasthttp.RequestCtx) {
-		tracing.ProcessSearchTracesRequest(ctx, 0)
+		tracinghandler.ProcessSearchTracesRequest(ctx, 0)
 	}
 }
